Agregue un menu para poder acceder a las distintas opciones. Agregue validaciones, para que solo los usuarios Administradores puedan ingresar a todas las paginas, y el resto solo a PostManager.

+++ SocialNetworkingSearchEngine/SocialNetWorkingSearchEngine/Controllers/QueryDefsController.cs
@@ -1,7 +1,5 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 using System.Web.Mvc;
 using Core.Domain;
 using Core.RepositoryInterfaces;
@@ -9,14 +7,21 @@ using DataAccess.DAO;
 using SocialNetWorkingSearchEngine.Models;
 
 namespace SocialNetWorkingSearchEngine.Controllers
-{   
+{
+    [Authorize]
     public class QueryDefsController : Controller
     {
-		private readonly IQueryDefRepository _querydefRepository;
+        private readonly IQueryDefRepository _querydefRepository;
+
+        public User Usuario
+        {
+            get { return (User)Session["Usuario"]; }
+            set { Session["Usuario"] = value; }
+        }
 
         public QueryDefsController(IQueryDefRepository querydefRepository)
         {
-			_querydefRepository = querydefRepository;
+            _querydefRepository = querydefRepository;
         }
 
         //
@@ -29,10 +34,16 @@ namespace SocialNetWorkingSearchEngine.Controllers
 
         public ViewResult Index(int page = 1, string like = "")
         {
-            IList<QueryDef> list = _querydefRepository.GetAll().Where(x => x.Query.StartsWith(like)).ToList();
-            var model = new CrudViewModel<QueryDef>(list, page, 10);
-            model.Filter = like;
-            return View(model);
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                IList<QueryDef> list = _querydefRepository.GetAll().Where(x => x.Query.StartsWith(like)).ToList();
+                var model = new CrudViewModel<QueryDef>(list, page, 10);
+                model.Filter = like;
+
+                return View(model); 
+            }
+
+            return View("LogOnUserControl");
         }
 
         //
@@ -40,7 +51,12 @@ namespace SocialNetWorkingSearchEngine.Controllers
 
         public ViewResult Details(int id)
         {
-            return View(_querydefRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_querydefRepository.GetById(id)); 
+            }
+
+            return View("LogOnUserControl");
         }
 
         //
@@ -48,8 +64,13 @@ namespace SocialNetWorkingSearchEngine.Controllers
 
         public ActionResult Create()
         {
-            return View();
-        } 
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(); 
+            }
+
+            return RedirectToAction("LogOn", "Account");
+        }
 
         //
         // POST: /QueryDefs/Create
@@ -57,20 +78,33 @@ namespace SocialNetWorkingSearchEngine.Controllers
         [HttpPost]
         public ActionResult Create(QueryDef querydef)
         {
-            if (ModelState.IsValid) {
-                _querydefRepository.Save(querydef);
-                return RedirectToAction("Index");
-            } else {
-				return View();
-			}
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                if (ModelState.IsValid)
+                {
+                    _querydefRepository.Save(querydef);
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View();
+                } 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
-        
+
         //
         // GET: /QueryDefs/Edit/5
- 
+
         public ActionResult Edit(int id)
         {
-             return View(_querydefRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_querydefRepository.GetById(id)); 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
@@ -79,21 +113,34 @@ namespace SocialNetWorkingSearchEngine.Controllers
         [HttpPost]
         public ActionResult Edit(QueryDef querydef)
         {
-            if (ModelState.IsValid) {
-				var entity = _querydefRepository.GetById(querydef.Id);
-				//_mapper.Map<QueryDef>(entity, querydef);
-                return RedirectToAction("Index");
-            } else {
-				return View();
-			}
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                if (ModelState.IsValid)
+                {
+                    var entity = _querydefRepository.GetById(querydef.Id);
+                    //_mapper.Map<QueryDef>(entity, querydef);
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    return View();
+                } 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
         // GET: /QueryDefs/Delete/5
- 
+
         public ActionResult Delete(int id)
         {
-            return View(_querydefRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_querydefRepository.GetById(id)); 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
@@ -102,9 +149,13 @@ namespace SocialNetWorkingSearchEngine.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-            _querydefRepository.Delete(_querydefRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                _querydefRepository.Delete(_querydefRepository.GetById(id));
+                return RedirectToAction("Index"); 
+            }
 
-            return RedirectToAction("Index");
+            return RedirectToAction("LogOn", "Account");
         }
     }
 }
