+++ SocialNetworkingSearchEngine/SocialNetWorkingSearchEngine/Controllers/WordsController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
@@ -7,19 +8,26 @@ using DataAccess.DAO;
 using SocialNetWorkingSearchEngine.Models;
 
 namespace SocialNetWorkingSearchEngine.Controllers
-{   
+{
+    [Authorize]
     public class WordsController : Controller
     {
-		private readonly IWordRepository _wordRepository;
+        private readonly IWordRepository _wordRepository;
+
+        public User Usuario
+        {
+            get { return (User)Session["Usuario"]; }
+            set { Session["Usuario"] = value; }
+        }
 
         public WordsController()
         {
-			_wordRepository = new WordRepository();
+            _wordRepository = new WordRepository();
         }
 
         public WordsController(IWordRepository wordRepository)
         {
-			_wordRepository = wordRepository;
+            _wordRepository = wordRepository;
         }
 
         //
@@ -27,18 +35,30 @@ namespace SocialNetWorkingSearchEngine.Controllers
 
         public ViewResult Index(int page = 1, string like = "")
         {
-            IList<Word> list = _wordRepository.GetAll().Where(x => x.Name.StartsWith(like)).ToList();
-            var model = new CrudViewModel<Word>(list, page, 20);
-            model.Filter = like;
-            return View(model);
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                IList<Word> list = _wordRepository.GetAll().Where(x => x.Name.StartsWith(like)).ToList();
+                var model = new CrudViewModel<Word>(list, page, 20);
+
+                model.Filter = like;
+
+                return View(model); 
+            }
+
+            return View("LogOnUserControl");
         }
 
         //
         // GET: /Words/Details/5
 
-        public ViewResult Details(System.Guid id)
+        public ViewResult Details(Guid id)
         {
-            return View(_wordRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_wordRepository.GetById(id)); 
+            }
+
+            return View("LogOnUserControl");
         }
 
         //
@@ -46,8 +66,13 @@ namespace SocialNetWorkingSearchEngine.Controllers
 
         public ActionResult Create()
         {
-            return View();
-        } 
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View();
+            }
+
+            return RedirectToAction("LogOn", "Account");
+        }
 
         //
         // POST: /Words/Create
@@ -55,19 +80,31 @@ namespace SocialNetWorkingSearchEngine.Controllers
         [HttpPost]
         public ActionResult Create(Word word)
         {
-            if (ModelState.IsValid) {
-                _wordRepository.Save(word);
-                return RedirectToAction("Index");
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                if (ModelState.IsValid)
+                {
+                    _wordRepository.Save(word);
+                    return RedirectToAction("Index");
+                }
+
+                return View();
             }
-            return View();
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
         // GET: /Words/Edit/5
- 
-        public ActionResult Edit(System.Guid id)
+
+        public ActionResult Edit(Guid id)
         {
-             return View(_wordRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_wordRepository.GetById(id));
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
@@ -76,30 +113,47 @@ namespace SocialNetWorkingSearchEngine.Controllers
         [HttpPost]
         public ActionResult Edit(Word word)
         {
-            if (ModelState.IsValid) {
-                _wordRepository.Save(word);
-                return RedirectToAction("Index");
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                if (ModelState.IsValid)
+                {
+                    _wordRepository.Save(word);
+                    return RedirectToAction("Index");
+                }
+
+                return View(); 
             }
-            return View();
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
         // GET: /Words/Delete/5
- 
-        public ActionResult Delete(System.Guid id)
+
+        public ActionResult Delete(Guid id)
         {
-            return View(_wordRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                return View(_wordRepository.GetById(id)); 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
 
         //
         // POST: /Words/Delete/5
 
         [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(System.Guid id)
+        public ActionResult DeleteConfirmed(Guid id)
         {
-            _wordRepository.Delete(_wordRepository.GetById(id));
+            if (Usuario != null && Usuario.IsAdmin)
+            {
+                _wordRepository.Delete(_wordRepository.GetById(id));
 
-            return RedirectToAction("Index");
+                return RedirectToAction("Index"); 
+            }
+
+            return RedirectToAction("LogOn", "Account");
         }
     }
 }
